requirements.txt
Flask
flask_cors
python-dotenv
deepgram-sdk
